pyproject.toml
[project]
name = "otel-langfuse"
version = "1.0.0"
description = "Teacher-Student Quiz System with LangGraph and OpenTelemetry"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "agentlightning>=0.3.0",  # AgentOps + OpenTelemetry 포함
    "poml>=0.0.8",
    "opentelemetry-instrumentation-langchain>=0.30.0",  # LangChain 자동 계측
    "pyyaml>=6.0",  # YAML 프롬프트 파일 로드
]

[project.scripts]
start = "uvicorn main:app --host 0.0.0.0 --port 8000"

[tool.uv]
dev-dependencies = []
